+++ tests/test-mathutil-gcd.c
@@ -5,15 +5,15 @@
 #include "utils.h"
 // -----------------------------------------------------------------------------
 int main(void) {
-  int trueResult = 7;
-  int a = 14, b = 35, gcd;
+  long trueResult = 7;
+  long a = 14, b = 35, gcd;
 
   // Get GCD
   calculateGCD(a, b, &gcd);
 
   // Check
   if (gcd != trueResult)
-    printf("Error:\n\tComputed: %d != True: %d\n", gcd, trueResult);
+    printf("Error:\n\tComputed: %ld != True: %ld\n", gcd, trueResult);
 
   return 0;
 }
